feat: modify delete-expense-function

+++ src/components/ExpensesList/ExpensesList.jsx
@@ -15,13 +15,33 @@ const ExpensesList = ({ selectedCategoryId, params }) => {
     navigate(`/expenses/${expenseId}/edit`, { state: { params, expenseId } })
   }
   const handleDeleteExpense = async (expenseId) => {
-    const { success, deletedExpenses } = await deleteExpense(expenseId)
-    if (success) {
-      noty('Delete Expense Successfully!', 'success')
-      console.log('deletedExpenses: ', deletedExpenses)
-      setExpensesData(prevExpenses => prevExpenses.filter(expense => expense.id !== expenseId))
-    } else {
-      noty('Delete Expense Failed!', 'error')
+    const expenseToDelete = expensesData.find(expense => expense.id === expenseId)
+    if (!expenseToDelete) {
+      console.error('Expense not found')
+      return
+    }
+    const groupInfo = expenseToDelete.group
+    const confirmMessage = groupInfo
+      ? 'Are you sure you want to delete this expense? (including the same group expenses after this)'
+      : 'Are you sure you want to delete this expense?'
+
+    const confirmDelete = window.confirm(confirmMessage)
+    if (confirmDelete) {
+      const { success, deletedExpenses } = await deleteExpense(expenseId)
+      if (success) {
+        noty('Delete Expense Successfully!', 'success')
+        if (Array.isArray(deletedExpenses)) {
+          deletedExpenses.forEach((expense, i) => {
+            noty(`Deleted expense ${i + 1}: Date - ${expense.date.split('T')[0]}, Name - ${expense.name}`, 'information')
+          })
+        } else if (typeof deletedExpenses === 'object' && deletedExpenses !== null) {
+          noty(`Deleted expense: Date - ${deletedExpenses.date.split('T')[0]}, Name - ${deletedExpenses.name}`, 'information')
+        }
+        console.log('deletedExpenses: ', deletedExpenses)
+        setExpensesData(prevExpenses => prevExpenses.filter(expense => expense.id !== expenseId))
+      } else {
+        noty('Delete Expense Failed!', 'error')
+      }
     }
   }
   const handleAddNewExpenses = () => {
